+++ requirements.txt
@@ -18,3 +18,5 @@ types-requests==2.27.27
 requests==2.26.0
 marshmallow-enum==1.5.1
 webargs==8.1.0
+python-jose==3.3.0
+python-jose-cryptodome==1.3.2
