+++ shweb/services/rest/__init__.py
@@ -9,7 +9,7 @@ from shweb.config import Config
 from shweb.services.rest.translate_helpers import get_locale
 from shweb.services.rest.router import register_blueprints
 from shweb.services.rest.errorhandlers import register_errorhandlers
-from shweb.services.rest.rest_helpers import utility_processor
+from shweb.services.rest.rest_helpers.common import utility_processor
 
 
 def create_app():
