+++ shweb/services/rest/routes/admin/routes/admin_index.py
@@ -4,7 +4,7 @@ from flask import render_template, make_response, current_app
 from flask_restful import Resource, reqparse, abort, request
 from flask_mobility.decorators import mobile_template
 
-from shweb.services.rest.routes.admin.auth import auth_required
+from shweb.services.rest.rest_helpers.common import auth_required
 from shweb.utils import upload_json, create_invalidation
