+++ shweb/services/rest/routes/public/feed.py
@@ -1,7 +1,7 @@
 """Рендер страницы новостей"""
 from flask import Blueprint, render_template
 
-from shweb.services.rest.rest_helpers import mobile_template
+from shweb.services.rest.rest_helpers.mobile import mobile_template
 
 
 blueprint = Blueprint("feed", __name__)
