--- shweb/util/enums.py
@@ -1,7 +1,0 @@
-import enum
-
-
-class ReleaseType(enum.Enum):
-    SINGLE = "Single"
-    ALBUM = "Album"
-    EP = "EP"
